requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# NLP and Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web App
streamlit>=1.25.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# LLM Integration
groq>=0.4.0

# Hyperparameter Optimization
optuna>=3.0.0

